web/backend/requirements-auth.txt
# Additional dependencies for authentication and integration
# Add to existing requirements.txt

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
email-validator>=2.0.0

# Database ORM
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0

# Background tasks
celery>=5.3.0
redis>=5.0.0

# File handling
python-magic>=0.4.27
aiofiles>=23.2.1

# Integration with existing codebase
# (These are already installed for the desktop app)
# opencv-python, numpy, torch, ultralytics, etc.
